refactor project files structure
