+++ client/index.js
@@ -1,92 +1,7 @@
-import { updateBrowsersStats, updateGlobalCoverageBar, updateToolsVersions, hideStatsPlaceholder } from './view/showStats.js';
-
-const API_HOST = 'http://localhost:5000/api';
-
-const form = document.getElementById('query_form');
-const textarea = document.getElementById('query_text_area');
-const errorMessage = document.getElementById('error_message');
-
-function initForm() {
-  form.addEventListener('submit', async (e) => {
-    if(!form.checkValidity()) {
-      return;
-    }
-    const formData = new FormData(form);
-    const query = formData.get('query')
-    e.preventDefault();
-    form.classList.add('Form--justSend');
-    textarea.addEventListener('input', () => {
-      form.classList.remove('Form--justSend');
-    }, {
-      once: true,
-    })
-    updateStatsView(query);
-  })
-  textarea.addEventListener('keypress', (e) => {
-    if (e.keyCode === 13 && !e.shiftKey) {
-      e.preventDefault();
-      form.dispatchEvent(new Event('submit'));
-    }
-  })
-}
+import { initForm } from './view/Form/form.js';
+import { initQueryLinks } from './view/QueryLink/queryLink.js';
 
 initForm();
-
-async function updateStatsView(query) {
-  let response;
-  try {
-    response = await fetch(`${API_HOST}?q=${encodeURIComponent(query)}`)
-    // TODO add loader
-  } catch (error) {
-    // TODO handle error
-    return false;
-  }
-
-  const data = await response.json();
-
-  if (!response.ok) {
-    if (data.message === 'Custom usage statistics was not provided') {
-      renderError(`This website does not support in my stats queries yet. Run Browserslist
- <a href="https://github.com/browserslist/browserslist#custom-usage-data" class="Link">locally</a>.`)
-      return false;
-    }
-    renderError(data.message);
-    return false;
-  }
-
-  const {
-    browsers,
-    versions,
-  } = data;
-
-  hideStatsPlaceholder();
-  updateBrowsersStats(browsers);
-  updateGlobalCoverageBar(browsers);
-  updateToolsVersions(versions);
-
-  return true;
-}
-
-function renderError(message) {
-  errorMessage.innerHTML = message;
-  form.classList.add('Form--serverError');
-  textarea.addEventListener('input', () => {
-    form.classList.remove('Form--serverError');
-  }, {
-    once: true,
-  })
-}
-
-function initQueryLinks() {
-  const links = document.querySelectorAll('a[data-query]');
-  links.forEach((item) => {
-    item.addEventListener('click', (e) => {
-      e.preventDefault();
-      textarea.value = item.getAttribute('data-query');
-      form.dispatchEvent(new Event('submit'));
-    })
-  })
-}
 
 initQueryLinks();
