+++ client/view/BrowserStats/BrowsersStat.css
